requirements-server.txt
beautifulsoup4==4.8.2
xmltodict==0.12.0
docutils==0.21
sphinx==8.2
sphinx-revealjs==3.2.0
pylint==2.5.3
Flask==2.0.3
Werkzeug>=2.2,<3.0
Flask-AutoIndex==0.6.6
gunicorn==20.1.0
sphinxcontrib.jquery==4.1
girth==0.8.0
numpy==1.22.3
pandas==1.4.2
mysql-connector-python==8.0.29
